msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>

	modified:   drivers/gpu/msm/adreno.c
	modified:   drivers/gpu/msm/adreno.h
	modified:   drivers/gpu/msm/adreno_dispatch.c
	modified:   drivers/gpu/msm/adreno_sysfs.c
	modified:   drivers/gpu/msm/kgsl_device.h
	modified:   drivers/gpu/msm/kgsl_ioctl.c