msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>

Conflicts:
	drivers/gpu/msm/kgsl_pwrscale.c

	modified:   drivers/gpu/msm/kgsl_pwrctrl.c
	modified:   drivers/gpu/msm/kgsl_pwrctrl.h
	modified:   drivers/gpu/msm/kgsl_pwrscale.c
	modified:   drivers/gpu/msm/kgsl_pwrscale.h
	modified:   drivers/gpu/msm/kgsl_trace.h